Simplified algorithm code files
