--- ini-rw/algorithms/IniKeyAlgorithms.hpp
@@ -1,18 +1,0 @@
-// File Name:     IniKeyAlgorithms.hpp
-// By:            Darian Benam (GitHub: https://github.com/BeardedFish/)
-// Date:          Monday, August 31, 2020
-
-#pragma once
-
-#include "../entities/IniKey.hpp"
-#include <string>
-#include <vector>
-
-namespace IniRW
-{
-	bool IsValidIniKey(const std::string& str);
-
-	IniKey* FindKey(std::vector<IniEntity*>& iniContents, const std::string& sectionName, const std::string& keyName);
-
-	std::string LowercaseValue(const std::string& str);
-}
