+++ ini-rw/algorithms/Validation.cpp
@@ -1,0 +1,34 @@
+// File Name:     Validation.cpp
+// By:            Darian Benam (GitHub: https://github.com/BeardedFish/)
+// Date:          Wednesday, September 2, 2020
+
+#include "Validation.hpp"
+
+namespace IniRW
+{
+	bool IsValidIniComment(const std::vector<char>& commentPrefixes, const std::string& str)
+	{
+		if (str.length() >= 1)
+		{
+			for (size_t i = 0; i < commentPrefixes.size(); i++)
+			{
+				if (str[0] == commentPrefixes[i])
+				{
+					return true;
+				}
+			}
+		}
+
+		return false;
+	}
+
+	bool IsValidIniKey(const std::string& str)
+	{
+		return str.length() >= 3 && str.find_first_of('=') != std::string::npos;
+	}
+
+	bool IsValidIniSection(const std::string& str)
+	{
+		return str.length() >= 2 && str[0] == SECTION_BEGINNING_CHAR && str[str.length() - 1] == SECTION_ENDING_CHAR;
+	}
+}
